Modificación en el formulario de la act 2 del ADC: Cambio de combo a radio para mostrar el TipoADC

+++ SistemaCenagas/SistemaCenagas/Consultas.cs
@@ -193,7 +193,7 @@ namespace SistemaCenagas
                     residencia = s.First().residencia,
                     id_proyecto = s.First().id_proyecto,
                     proyecto = s.First().proyecto,
-                    avance_ADC = s.Sum(a => a.avance_ADC) / s.Count(),
+                    avance_ADC = (int)(s.Sum(a => a.avance_ADC) / s.Count()),
                     avance_Pre = 0,
                     avance_Fisico = 0
                 }).ToList();
